[Project P1] Extract GalaxyScene.tsx and create layer-constants.ts
Task 1.1: Split GalaxyRing3D.tsx (1200 lines) into:
- GalaxyScene.tsx (1158 lines): all 3D scene components, terrain,
  entities, effects, helpers
- GalaxyRing3D.tsx (60 lines): thin Canvas wrapper only

Task 1.2: Create layer-constants.ts with named z-index constants
for the full UI stack (LAYER_3D_DEFAULT, LAYER_3D_INTERACTIVE,
LAYER_GAME_UI, LAYER_TD_PANEL, etc.). Replace inline magic numbers
in GalaxyRing3D.tsx and document mapping in GalaxyTDPanel.module.css.
