+++ src/components/rhythmia/tetris/components/GalaxyTDPanel.module.css
@@ -4,7 +4,7 @@
     position: fixed;
     inset: 0;
     pointer-events: none;
-    z-index: 15;
+    z-index: 15; /* LAYER_TD_PANEL */
     font-family: 'Inter', sans-serif;
 }
